requirements.txt
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
# google_workspace_utils @ git+https://gitlab.com/firology/google-workspace-utils.git
google_workspace_utils >= 0.1.1
flask
Pillow
